Résolution d'une erreur qui faisait crash le bot

+++ events/messageCreate.js
@@ -1,4 +1,5 @@
 const {Events} = require('discord.js');
+const {log} = require("../utils/utils");
 
 module.exports = {
   name: Events.MessageCreate,
@@ -19,12 +20,16 @@ module.exports = {
     } else {
       // Ignore all bot messages
       if (message.author.bot) return;
+      if (!message.guild) {
+        message.reply('Je ne suis pas encore prêt pour les messages privés, désolé !');
+        log(message.content);
+      } else {
+        const {attribution_sports} = require(`../serveur/channels/channels_${message.guild.id}.json`);
 
-      const {attribution_sports} = require(`../serveur/channels/channels_${message.guild.id}.json`);
-
-      // Ignore all messages not in the attribution_sports channel
-      if (message.channelId === attribution_sports) {
-        message.delete();
+        // Ignore all messages not in the attribution_sports channel
+        if (message.channelId === attribution_sports) {
+          message.delete();
+        }
       }
     }
   },
